pos-graduacao/02-dw-modeling-implementation-governance/16-governance-and-dq-with-great-expectations/Lab4/requirements.txt
great-expectations==0.18.16
psycopg2==2.9.9
pyyaml==6.0
sqlalchemy==1.4.39